test(plugins): isolate bundle inventory coverage
