+++ src/plugins/status.registry-snapshot.test.ts
@@ -26,7 +26,6 @@ import {
   isColdPluginRuntimeLoaded,
 } from "./test-helpers/cold-plugin-fixtures.js";
 import { cleanupTrackedTempDirs, makeTrackedTempDir } from "./test-helpers/fs-fixtures.js";
-import { createBundleInstallFixtureFactory } from "./test-helpers/install-fixtures.js";
 import { writeManagedNpmPlugin } from "./test-helpers/managed-npm-plugin.js";
 
 const tempDirs: string[] = [];
@@ -34,8 +33,6 @@ const tempDirs: string[] = [];
 function makeTempDir() {
   return makeTrackedTempDir("openclaw-plugin-status", tempDirs);
 }
-
-const setupBundleInstallFixture = createBundleInstallFixtureFactory(makeTempDir);
 
 function createWorkspacePluginFixture(workspaceDir: string, pluginId: string) {
   const rootDir = path.join(workspaceDir, ".openclaw", "extensions", pluginId);
@@ -99,61 +96,6 @@ function expectFields(actual: Record<string, unknown>, expected: Record<string, 
 }
 
 describe("buildPluginRegistrySnapshotReport", () => {
-  it.each([
-    {
-      bundleFormat: "agent",
-      enabled: true,
-      registrySource: "derived",
-      capabilities: ["skills"],
-    },
-    {
-      bundleFormat: "claude",
-      enabled: false,
-      registrySource: "persisted",
-      capabilities: ["skills"],
-    },
-    {
-      bundleFormat: "cursor",
-      enabled: true,
-      registrySource: "persisted",
-      capabilities: ["skills", "commands"],
-    },
-  ] as const)(
-    "preserves $bundleFormat bundle capabilities in $registrySource inventory (enabled=$enabled)",
-    async ({ bundleFormat, enabled, registrySource, capabilities }) => {
-      const name = `${bundleFormat}-capability-fixture`;
-      const { pluginDir, extensionsDir } = setupBundleInstallFixture({ bundleFormat, name });
-      const stateDir = path.dirname(extensionsDir);
-      const workspaceDir = path.dirname(stateDir);
-      const params = {
-        config: {
-          plugins: {
-            load: { paths: [pluginDir] },
-            entries: { [name]: { enabled } },
-          },
-        },
-        workspaceDir,
-        env: {
-          ...createColdPluginHermeticEnv(workspaceDir, { bundledPluginsDir: makeTempDir() }),
-          OPENCLAW_STATE_DIR: stateDir,
-        },
-      };
-      if (registrySource === "persisted") {
-        await refreshPluginRegistry({ ...params, stateDir, reason: "manual" });
-      }
-
-      const report = buildPluginRegistrySnapshotReport(params);
-
-      expectFields(requirePlugin(report.plugins, name), {
-        format: "bundle",
-        bundleFormat,
-        bundleCapabilities: capabilities,
-        enabled,
-      });
-      expect(report.registrySource).toBe(registrySource);
-    },
-  );
-
   it("uses the configured system owner for ambient plugin inventory", () => {
     const tempRoot = makeTempDir();
     const mainWorkspace = path.join(tempRoot, "main-workspace");
